добавит 3 файла в mock (вместо одного). рефакторинг create-point-view
