+++ src/mock/destinations.js
@@ -1,0 +1,61 @@
+const mockDestinations = [
+  {
+    id: 1,
+    description: 'Chamonix, is a beautiful city, a true asian pearl, with crowded streets.',
+    name: 'Chamonix',
+    pictures: [
+      {
+        src: 'https://loremflickr.com/248/152?random=5',
+        description: 'Chamonix parliament building'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=18',
+        description: 'Chamonix 2'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=21',
+        description: 'Chamonix 3'
+      }
+    ]
+  },
+  {
+    id: 2,
+    description: 'Amsterdam is the capital and most populated city of the Netherlands.',
+    name: 'Amsterdam',
+    pictures: [
+      {
+        src: 'https://loremflickr.com/248/152?random=6',
+        description: 'van gogh museum'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=555',
+        description: 'Amsterdam 2'
+      },
+    ]
+  },
+  {
+    id: 3,
+    description: 'Geneva is the second-most populous city in Switzerland (after Zürich) and the most populous of the French-speaking Romandy.',
+    name: 'Geneva',
+    pictures: [
+      {
+        src: 'https://loremflickr.com/248/152?random=7',
+        description: 'Geneva Cathedral'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=13',
+        description: 'Geneva 2'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=28',
+        description: 'Geneva 3'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=29',
+        description: 'Geneva 4'
+      }
+    ]
+  }
+];
+
+export {mockDestinations};
