+++ src/mock/offers.js
@@ -1,0 +1,122 @@
+const mockOffers = [
+  {
+    type: 'Taxi',
+    offers: [
+      {
+        id: 'taxi-1',
+        title: 'Order Uber',
+        price: 120
+      },
+      {
+        id: 'taxi-2',
+        title: 'Select music',
+        price: 50
+      }
+    ]
+  },
+  {
+    type: 'Bus',
+    offers: [
+      {
+        id: 'bus-1',
+        title: 'add meal',
+        price: 30
+      }
+    ]
+  },{
+    type: 'Train',
+    offers: [
+      {
+        id: 'train-1',
+        title: 'Choose a seat',
+        price: 50
+      }
+    ]
+  },{
+    type: 'Ship',
+    offers: [
+      {
+        id: 'ship-1',
+        title: 'Choose a seat',
+        price: 20
+      }
+    ]
+  },{
+    type: 'Drive',
+    offers: [
+      {
+        id: 'drive-1',
+        title: 'Rent a car',
+        price: 1000
+      }
+    ]
+  },{
+    type: 'Flight',
+    offers: [
+      {
+        id: 'flight-1',
+        title: 'Add luggage',
+        price: 120
+      },
+      {
+        id: 'flight-2',
+        title: 'Switch to comfort',
+        price: 50
+      },
+      {
+        id: 'flight-3',
+        title: 'Add meal',
+        price: 30
+      },
+      {
+        id: 'flight-4',
+        title: 'Choose seats',
+        price: 150
+      },
+    ]
+  },{
+    type: 'Check-in',
+    offers: [
+      {
+        id: 'check-in-1',
+        title: 'Add breakfast',
+        price: 300
+      },
+      {
+        id: 'check-in-2',
+        title: 'Add dinner',
+        price: 350
+      },
+      {
+        id: 'check-in-3',
+        title: 'All inclusive',
+        price: 700
+      }
+    ]
+  },{
+    type: 'Sightseeing',
+    offers: [
+      {
+        id: 'sightseeing-1',
+        title: 'Book tickets',
+        price: 120
+      },
+      {
+        id: 'sightseeing-2',
+        title: 'Lunch in city',
+        price: 50
+      }
+    ]
+  },{
+    type: 'Restaurant',
+    offers: [
+      {
+        id: 'restaurant-1',
+        title: 'meet the chef',
+        price: 120
+      }
+    ]
+  },
+];
+
+export {mockOffers};
