+++ src/model/points-model.js
@@ -1,11 +1,24 @@
 import {getRandomPoint} from '../mock/point.js';
+import { mockOffers } from '../mock/offers.js';
+import { mockDestinations } from '../mock/destinations.js';
 
 const POINT_COUNT = 4;
 
 export default class PointsModel {
   points = Array.from({length: POINT_COUNT}, getRandomPoint);
+  offers = mockOffers;
+  destinations = mockDestinations;
+
 
   getPoints() {
     return this.points;
+  }
+
+  getOffers() {
+    return this.offers;
+  }
+
+  getDestinations() {
+    return this.destinations;
   }
 }
